Реализованы авторизация и REST API для работы с пользователями

+++ src/main/java/by/speedteam/speedwagon/controllers/UserController.java
@@ -1,8 +1,10 @@
 package by.speedteam.speedwagon.controllers;
 
 import by.speedteam.speedwagon.models.User;
+import by.speedteam.speedwagon.payload.requests.users.UpdateUserRequest;
 import by.speedteam.speedwagon.payload.responses.ErrorResponse;
 import by.speedteam.speedwagon.payload.responses.SuccessResponse;
+import by.speedteam.speedwagon.payload.responses.users.UserDto;
 import by.speedteam.speedwagon.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -10,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class UserController {
@@ -23,7 +26,7 @@ public class UserController {
     @GetMapping("/api/v1/users")
     public ResponseEntity<?> getAllUsers() {
         try {
-            List<User> users = userService.getAllUsers();
+            List<UserDto> users = userService.getAllUsers();
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new SuccessResponse<>("Users retrieved successfully", users));
         } catch (Exception e) {
@@ -36,8 +39,7 @@ public class UserController {
     @GetMapping("/api/v1/users/{id}")
     public ResponseEntity<?> getUserById(@PathVariable long id) {
         try {
-            User user = userService.getUserById(id)
-                    .orElseThrow(() -> new RuntimeException("User not found"));
+            UserDto user = userService.getUserDtoById(id);
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new SuccessResponse<>("User retrieved successfully", user));
         } catch (RuntimeException e) {
@@ -50,28 +52,25 @@ public class UserController {
         }
     }
 
-//    @PostMapping("/users")
-//    public ResponseEntity<?> createUser(@RequestBody User user) {
-//        try {
-//            User newUser = userService.createUser(user);
-//            return ResponseEntity.status(HttpStatus.CREATED)
-//                    .body(new SuccessResponse<>("User created successfully", newUser));
-//        } catch (Exception e) {
-//            System.err.println(e.getMessage());
-//            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-//                    .body(new ErrorResponse("Internal server error"));
-//        }
-//    }
-
     @PutMapping("/api/v1/users/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable long id, @RequestBody User user) {
+    public ResponseEntity<?> updateUser(@PathVariable long id, @RequestBody UpdateUserRequest updateUserRequest) {
+        Optional<User> userData = userService.getUserById(id);
+        if (userData.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ErrorResponse("User not found"));
+        }
+        if (userService.isEmailAlreadyExists(updateUserRequest.getEmail())) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ErrorResponse("Email already exists"));
+        }
+        if (userService.isPhoneAlreadyExists(updateUserRequest.getPhone())) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new ErrorResponse("Phone already exists"));
+        }
         try {
-            User updatedUser = userService.updateUser(id, user);
+            UserDto updatedUser = userService.updateUser(id, updateUserRequest);
             return ResponseEntity.status(HttpStatus.OK)
                     .body(new SuccessResponse<>("User updated successfully", updatedUser));
-        } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new ErrorResponse(e.getMessage()));
         } catch (Exception e) {
             System.err.println(e.getMessage());
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -81,15 +80,16 @@ public class UserController {
 
     @DeleteMapping("/api/v1/users/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable long id) {
+        Optional<User> userData = userService.getUserById(id);
+        if (userData.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ErrorResponse("User not found"));
+        }
         try {
             userService.deleteUser(id);
             return ResponseEntity.status(HttpStatus.OK)
-                    .body(new SuccessResponse<>("User deleted successfully", null));
-        } catch (RuntimeException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(new ErrorResponse(e.getMessage()));
+                    .body(new SuccessResponse<>("User deleted successfully", id));
         } catch (Exception e) {
-            System.err.println(e.getMessage());
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body(new ErrorResponse("Internal server error"));
         }
